requirements.txt
annotated-types==0.7.0
anyio==4.12.0
captcha==0.7.1
certifi==2025.11.12
click==8.3.1
fastapi==0.109.2
filetype==1.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.26.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.3
MarkupSafe==3.0.3
numpy==1.26.4
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
pydantic==2.12.5
pydantic-settings==2.2.1
pydantic_core==2.41.5
pytest==8.0.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.36.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.27.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
